pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "render_engine"
dynamic = ["version"]
description = "A Flexible Static Site Generator for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitpython",
    "jinja2",
    "markdown2",
    "more-itertools",
    "pluggy",
    "python-dateutil",
    "python-frontmatter",
    "python-slugify",
    "render-engine-parser",
    "render-engine-markdown",
    "tzdata",
]

[project.optional-dependencies]
cli = [
    "cookiecutter",
    "rich",
    "typer",
    "watchfiles",
]
extras = [
  "render-engine-sitemap"
]
dev = [
    "cookiecutter",
    "rich",
    "typer",
    "watchfiles",
    "httpx",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "pymdown-extensions",
    "ruff",
    "pre-commit",
    "mypy"
    ]


[tool.setuptools_scm]
local_scheme = "no-local-version"
# version_scheme = "python-simplified-semver"

[project.urls]
homepage = "https://github.com/kjaymiller/render_engine/"
repository = "https://github.com/kjaymiller/render_engine/"
documentation = "https://render-engine.readthedocs.io/en/latest/"

[tool.setuptools.package-data]
"*" = ["*.html", "*.txt", "*.xml", "*.md"]

[project.scripts]
render-engine = "render_engine.cli.cli:cli"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = ["--cov=src", "--cov-report=term-missing", "-ra", "-q"]

[tool.semantic_release]
version_toml = "pyproject.toml:project.version"
branch = "main"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
